Task 2/requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
flask>=2.3.0
pillow>=10.0.0
opencv-python>=4.8.0